+++ iterative/resource_runner.go
@@ -380,7 +380,7 @@ func provisionerCode(d *schema.ResourceData) (string, error) {
 				Attributes: AttributesType{
 					ID:                 d.Id(),
 					Cloud:              d.Get("cloud").(string),
-					Region:             d.Get("region").(string),
+					Region:             utils.StripAvailabilityZone(d.Get("region").(string)),
 					Name:               d.Get("name").(string),
 					Labels:             "",
 					IdleTimeout:        d.Get("idle_timeout").(int),
