+++ iterative/utils/helpers.go
@@ -23,3 +23,13 @@ func SetId(d *schema.ResourceData) {
 		}
 	}
 }
+
+func StripAvailabilityZone(region string) string {
+	lastChar := region[len(region)-1]
+	// 0x61(a) to 0x7a(z)
+	if lastChar >= 0x61 && lastChar <= 0x71 {
+		return region[:len(region)-1]
+	} else {
+		return region
+	}
+}
